feature[feature/tattoo-artists-advantagesTRS]:responsivess and mobile styled
