+++ client/src/assets/index.ts
@@ -1,6 +1,6 @@
 import LogoCITi from './logo.png';
-import Foto1 from './Ilustração=One.png';
-import Foto2 from './Ilustração=Two.png';
+import Foto1 from './Ilustração=One.svg';
+import Foto2 from './Ilustração=Two.svg';
 import Foto3 from './Ilustração=Three.svg';
 import NoirProMediumItalicWoff2 from './noir-pro/NoirPro-Webfont/NoirPro-MediumItalic.woff2';
 import NoirProMediumItalicWoff from './noir-pro/NoirPro-Webfont/NoirPro-MediumItalic.woff';
