+++ client/src/components/index.tsx
@@ -1,5 +1,5 @@
 import { Card } from './Card';
 
 export {
-    Card
+    Card,
 };
